Move get_tokenizer to utils
