+++ torchtext/data/field.py
@@ -5,6 +5,7 @@ from torch.autograd import Variable
 
 from .dataset import Dataset
 from .pipeline import Pipeline
+from .utils import get_tokenizer
 from ..vocab import Vocab
 
 
@@ -186,25 +187,3 @@ class Field(object):
         if self.include_lengths:
             return Variable(arr, volatile=not train), lengths
         return Variable(arr, volatile=not train)
-
-
-def get_tokenizer(tokenizer):
-    if callable(tokenizer):
-        return tokenizer
-    if tokenizer == 'spacy':
-        try:
-            import spacy
-            spacy_en = spacy.load('en')
-            return lambda s: [tok.text for tok in spacy_en.tokenizer(s)]
-        except ImportError:
-            print("Please install SpaCy and the SpaCy English tokenizer. "
-                  "See the docs at https://spacy.io for more information.")
-            raise
-        except AttributeError:
-            print("Please install SpaCy and the SpaCy English tokenizer. "
-                  "See the docs at https://spacy.io for more information.")
-            raise
-    raise ValueError("Requested tokenizer {}, valid choices are a "
-                     "callable that takes a single string as input "
-                     "and \"spacy\" for the SpaCy English "
-                     "tokenizer.".format(tokenizer))
